Google Pay Functional, Code Layout Fixed
