+++ app/src/main/java/com/example/rusticroots/MainActivity.kt
@@ -5,29 +5,16 @@ import android.util.Log
 import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.result.ActivityResult
-import androidx.activity.result.IntentSenderRequest
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.defaultMinSize
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.Observer
+import com.example.rusticroots.ui.modules.GooglePayButton
 import com.example.rusticroots.ui.theme.RusticRootsTheme
 import com.example.rusticroots.viewmodel.PaymentGViewModel
-import com.google.android.gms.common.api.ApiException
-import com.google.android.gms.common.api.CommonStatusCodes
-import com.google.android.gms.common.api.ResolvableApiException
-import com.google.android.gms.wallet.PaymentData
-import org.json.JSONException
-import org.json.JSONObject
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -42,7 +29,7 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colors.background
                 ) {
-                    GooglePayButton(paymentGVM = paymentGVM)
+                    MyApp(paymentGVM = paymentGVM)
                 }
             }
         }
@@ -64,115 +51,11 @@ class MainActivity : ComponentActivity() {
             Log.e("**********************","NOT AVAILABLE")
         }
     }
-
-    fun requestPayment(paymentGVM: PaymentGViewModel){
-
-        val dummyPriceCents = 100L
-        val shippingCostCents = 900L
-        val task = paymentGVM.getLoadPaymentDataTask(dummyPriceCents + shippingCostCents)
-
-        task.addOnCompleteListener { completedTask ->
-            if (completedTask.isSuccessful) {
-                completedTask.result.let(::handlePaymentSuccess)
-                Log.d("*************************","completedTask.isSuccessful")
-            } else {
-                Log.e("*************************","completedTask.isNOTSuccessful")
-                when (val exception = completedTask.exception) {
-                    is ResolvableApiException -> {
-                        Log.d("*************************","ResolvableApiException")
-                        resolvePaymentForResult.launch(
-                            IntentSenderRequest.Builder(exception.resolution).build()
-                        )
-                    }
-                    is ApiException -> {
-                        handleError(exception.statusCode, exception.message)
-                    }
-                    else -> {
-                        handleError(
-                            CommonStatusCodes.INTERNAL_ERROR, "Unexpected non API" +
-                                    " exception when trying to deliver the task result to an activity!"
-                        )
-                    }
-                }
-            }
-        }
-
-    }
-    private val resolvePaymentForResult = registerForActivityResult(ActivityResultContracts.StartIntentSenderForResult()) {
-            result: ActivityResult ->
-        when (result.resultCode) {
-            RESULT_OK ->
-                result.data?.let { intent ->
-                    PaymentData.getFromIntent(intent)?.let(::handlePaymentSuccess)
-                }
-
-            RESULT_CANCELED -> {
-                // The user cancelled the payment attempt
-            }
-        }
-    }
-    private fun handlePaymentSuccess(paymentData: PaymentData) {
-        val paymentInformation = paymentData.toJson()
-
-        try {
-            // Token will be null if PaymentDataRequest was not constructed using fromJson(String).
-            val paymentMethodData = JSONObject(paymentInformation).getJSONObject("paymentMethodData")
-            val billingName = paymentMethodData.getJSONObject("info")
-                .getJSONObject("billingAddress").getString("name")
-            Log.d("BillingName", billingName)
-
-            Toast.makeText(this, getString(R.string.payments_show_name, billingName), Toast.LENGTH_LONG).show()
-
-            // Logging token string.
-            Log.d("Google Pay token", paymentMethodData
-                .getJSONObject("tokenizationData")
-                .getString("token"))
-
-        } catch (error: JSONException) {
-            Log.e("handlePaymentSuccess", "Error: $error")
-        }
-    }
-
-    /**
-     * At this stage, the user has already seen a popup informing them an error occurred. Normally,
-     * only logging is required.
-     *
-     * @param statusCode will hold the value of any constant from CommonStatusCode or one of the
-     * WalletConstants.ERROR_CODE_* constants.
-     * @see [
-     * Wallet Constants Library](https://developers.google.com/android/reference/com/google/android/gms/wallet/WalletConstants.constant-summary)
-     */
-    private fun handleError(statusCode: Int, message: String?) {
-        Log.e("Google Pay API error", "Error code: $statusCode, Message: $message")
-    }
-
-    @Composable
-    fun GooglePayButton(paymentGVM: PaymentGViewModel) {
-        var enabled by remember { mutableStateOf(true) }
-
-        Button(
-            onClick = { requestPayment(paymentGVM) },
-            enabled = enabled,
-            modifier = Modifier
-                .defaultMinSize(minWidth = 152.dp),
-            colors = ButtonDefaults.buttonColors(backgroundColor = Color.White),
-        ) {
-            Image(painter = painterResource(
-                id = R.drawable.pay_with_googlepay_button_content),
-                contentDescription = "Google pay Button"
-            )
-        }
-    }
-
-
 }
 
-
-
-@Preview(showBackground = true)
 @Composable
-fun DefaultPreview() {
-    RusticRootsTheme {
-        //GooglePayButton()
+fun MyApp(paymentGVM: PaymentGViewModel) {
+    Column() {
+        GooglePayButton(paymentGVM = paymentGVM)
     }
 }
